backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.1

# Redis
redis[hiredis]==5.0.1

# Encryption
cryptography==42.0.2

# AI / LLM
anthropic==0.18.1
openai==1.12.0

# Vector DB
pinecone-client==3.1.0

# Telegram
pyrogram==2.0.106
tgcrypto==1.2.5

# HTTP Client
httpx==0.27.0
aiohttp==3.9.3

# Task Queue
celery[redis]==5.3.6

# File Processing
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2

# Media Processing
pydub==0.25.1
Pillow==10.2.0

# Auth
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# Utilities
tenacity==8.2.3
structlog==24.1.0
stripe==7.9.0
orjson==3.9.15
python-dateutil==2.8.2
pytz==2024.1
# Additional dependencies for local dev added by agent
email-validator==2.1.0.post1
sentry-sdk==1.40.6
